expose driver disabling preference to users
Each driver now has its own Boolean preference of the form
"disable<drivername>", with <drivername> taken from the key-name used
for the driver in "allDrivers" in lib/main.js. The preference indicates
whether the use of the HTML5 player should be disabled for that site.

This replaces the hidden preference of comma-separated list of drivers
which previously did not have a UI for enabling drivers once they were
disabled.